refactor(app): Refactored the code
I have made the changes as requested and then also this time made the LessonHeading -> Header, now a global components that is made accessible to everyone who would like to use them in any other components
