+++ src/components/Assignment/AssignmentBody.jsx
@@ -1,0 +1,11 @@
+import Assignment from './Assignment'
+
+const AssignmentBody = () => {
+  return (
+    <section>
+      <Assignment />
+    </section>
+  )
+}
+
+export default AssignmentBody
