--- src/components/Dashboards/AssignmentBody.jsx
@@ -1,15 +1,0 @@
-import Assignment from "./Assignment";
-
-const AssignmentBody = () => {
-    return (
-        <section>
-            <Assignment/>
-            <Assignment/>
-            <Assignment/>
-            <Assignment/>
-            <Assignment/>
-        </section>
-    )
-};
-
-export default AssignmentBody;
