+++ src/components/Header.jsx
@@ -1,9 +1,9 @@
-import clock from '../../assets/clock.png'
-import circled from '../../assets/circled.png'
+import clock from './../assets/clock.png'
+import circled from './../assets/circled.png'
 
-const LessonHeading = ({ title, para, stream, duration }) => {
+const Header = ({ title, para, stream, duration }) => {
   return (
-    <header className="flex flex-col md:flex-row md:items-center mt-10 justify-between border-b py-3 md:pb-4 md:gap-0 gap-4">
+    <header className="flex flex-col md:flex-row md:items-center mt-10 md:mt-0 justify-between border-b py-3 md:pb-4 md:gap-0 gap-4">
       <div>
         <h1 className="font-bold text-3xl">{title}</h1>
         <p className="text-[#444444]">{para}</p>
@@ -34,4 +34,4 @@ const LessonHeading = ({ title, para, stream, duration }) => {
   )
 }
 
-export default LessonHeading
+export default Header
