+++ src/pages/Assingments.jsx
@@ -1,11 +1,16 @@
-import AssignmentHeading from "../components/Dashboards/AssignmentHeading"
-import AssignmentBody from "../components/Dashboards/AssignmentBody";
+import AssignmentBody from '../components/Assignment/AssignmentBody'
+import Header from '../components/Header'
 
 const Assingments = () => {
   return (
     <section className="flex flex-col">
-        <AssignmentHeading/>
-        <AssignmentBody/>
+      <Header
+        title={'Your Assignments 🧑🏾‍💻'}
+        para={'Your assignment central – explore them all here.'}
+        stream={'Backend (Caramel) - JavaScript'}
+        duration={'90d'}
+      />
+      <AssignmentBody />
     </section>
   )
 }
