+++ src/pages/Lessons.jsx
@@ -1,10 +1,10 @@
 import LessonCard from '../components/Lessons/LessonCard'
-import LessonHeading from '../components/Lessons/LessonHeading'
+import Header from '../components/Header'
 
 const Lessons = () => {
   return (
     <div className="lessons ">
-      <LessonHeading
+      <Header
         title="Your Lessons"
         para="Explore all your lessons at a glance."
         stream="Backend (Caramel) - Javascript"
